Moved mock app store into tank test module
